Refactored AppConfig.py to app_config.py
